Implement console email backend and start to refactor views
