+++ server/views/pagination.py
@@ -1,0 +1,8 @@
+def paginate(page):
+    size = 20
+
+    start = (page - 1) * size
+
+    end = start + size
+
+    return start, end
