+++ tests/test_registration.py
@@ -1,4 +1,3 @@
-from unittest.mock import patch
 from http import HTTPStatus
 
 from server.models import VerificationEmail
@@ -7,12 +6,13 @@ from server.models import VerificationEmail
 def test_faculty_registration_email(client, requests_mock):
     email = 'test@hms.harvard.edu'
 
-    requests_mock.post(f'https://api.sparkpost.com/api/v1/transmissions', {}, reason='OK')
+    requests_mock.post(
+        f'https://api.sparkpost.com/api/v1/transmissions', {}, reason='OK'
+    )
 
-    with patch('server.emails.verify_email_configured'):
-        response = client.post(
-            '/api/send-faculty-verification-email', json={'email': email}
-        )
+    response = client.post(
+        '/api/send-faculty-verification-email', json={'email': email}
+    )
 
     assert response.json['email'] == email
 
@@ -24,11 +24,12 @@ def test_faculty_registration_email(client, requests_mock):
 
 
 def test_faculty_registration_invalid_email(client, requests_mock):
-    requests_mock.post(f'https://api.sparkpost.com/api/v1/transmissions', {}, reason='OK')
+    requests_mock.post(
+        f'https://api.sparkpost.com/api/v1/transmissions', {}, reason='OK'
+    )
 
-    with patch('server.emails.verify_email_configured'):
-        response = client.post(
-            '/api/send-faculty-verification-email', json={'email': 'test@test.com'}
-        )
+    response = client.post(
+        '/api/send-faculty-verification-email', json={'email': 'test@test.com'}
+    )
 
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY.value
